Add vertical spacing around list items in drink notes
This should impose vertical spacing between list items, regardless of
whether or not there is an empty line between list items in the data
from Contentful. Adding an empty line in Contentful actually wraps
the list item text in a paragraph tag, which comes with its own
margin. This solution works due to margin collapsing.