fix(dialog): avoid content jump when showing feedback form
@reach/dialog v0.2.0 (bb429a15fed88ff446bfa83174ebafb6f50953b9) introduced the use of
react-remove-scroll and react-remove-scroll-bar, which makes an opinionated choice to hide the
scroll bar and add margin to the body to offset the missing scroll bar width. This causes the page
content to jump upon opening the feedback form. Abusing CSS specificity to force our own overflow-y
and margin values remedies this until the use of react-remove-scroll is customizable in
@reach/dialog (if that ever even happens).